Improve plot of new URLs per crawl
Show only URLs in the plot of "new items" to make it more readable.
Numbers of new "pages" (aka. "captures") and "digests" are not
really informative: it's just the size of the crawls resp. a number
close to the size as there are very few exact duplicates with
identical content digests. Drop the two item types and keep only
the URLs.